+++ CHANGELOG.rst
@@ -378,3 +378,5 @@ Bug fixes
   `#585 <https://github.com/openego/eGon-data/issues/585>`_
 * Change source for H2 steel tank storage to Danish Energy Agency
   `#605 <https://github.com/openego/eGon-data/issues/605>`_
+* Create and fill transfer_busses table in substation-dataset
+  `#610 <https://github.com/openego/eGon-data/issues/610>`_